toolkit/content/tests/chrome/test_bug418874.xul

<?xml version="1.0"?>

<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is XUL Textbox Widget Test code
   -
   - The Initial Developer of the Original Code is
   - Graeme McCutcheon <graememcc_firefox@graeme-online.co.uk>.
   - Portions created by the Initial Developer are Copyright (C) 2008
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->


<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Textbox with placeholder test" width="500" height="600"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <hbox>
    <textbox id="t1" placeholder="empty"/>
  </hbox>

  <hbox>
    <textbox id="t2" placeholder="empty"/>
  </hbox>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml" style="height: 300px; overflow: auto;">
    <p id="display">
    </p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
    SimpleTest.waitForExplicitFinish();

    function doTest() {
      var t1 = $("t1");
      var t2 = $("t2");
      setTextboxValue(t1, "1");
      var t1Enabled = {};
      var t1CanUndo = {};
      t1.editor.canUndo(t1Enabled, t1CanUndo);
      is(t1CanUndo.value, true,
         "undo correctly enabled when placeholder was not changed through property");

      t2.placeholder = "reallyempty";
      setTextboxValue(t2, "2");
      var t2Enabled = {};
      var t2CanUndo = {};
      t2.editor.canUndo(t2Enabled, t2CanUndo);
      is(t2CanUndo.value, true,
         "undo correctly enabled when placeholder explicitly changed through property");

      SimpleTest.finish();
   }

   function setTextboxValue(textbox, value) {
      textbox.focus();
      for (var i = 0; i < value.length; ++i) {
          synthesizeKey(value.charAt(i), {});
      }
      textbox.blur();
   }

   SimpleTest.waitForFocus(doTest);

  ]]></script>

</window>
